tests/srp/rc: Increase maximum request size
A new feature for the SRP initiator and target drivers in kernel v5.0-rc1
is that both the initiator and target drivers support immediate data.
Since the default maximum immediate data size is 8 KB, increase the target
maximum request size such that the SRP initiator can send write requests
with up to 8 KB of immediate data.

Signed-off-by: Bart Van Assche <bvanassche@acm.org>